requirements.txt
Django==1.9.7
Pillow==3.2.0
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
djangorestframework==3.3.3
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
wsgiref==0.1.2
